refactor: LottoBuyPrice의 허용 금액 단위 상수 타입을 실수형에서 정수형으로 변경
기존 상수 타입을 실수형으로 설정할 필요성이 없다고 판단되어 정수형으로 변경함.

+++ src/main/java/lotto/domain/LottoBuyPrice.java
@@ -2,7 +2,7 @@ package lotto.domain;
 
 public class LottoBuyPrice {
 
-    private static final double ALLOWED_AMOUNT_UNIT = 1000.0;
+    private static final int ALLOWED_AMOUNT_UNIT = 1000;
 
     private final int price;
